backend/requirements.txt
about-time==3.1.1
alive-progress==2.3.1
attrs==19.3.0
Automat==0.8.0
bcrypt==3.2.0
blinker==1.4
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
click==8.0.4
cloud-init==19.1
cmdstanpy==0.9.68
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
convertdate==2.4.0
cryptography==2.8
cycler==0.11.0
Cython==0.29.28
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
ephem==4.1.3
filelock==3.6.0
Flask==2.0.3
Flask-WTF==1.0.0
fonttools==4.30.0
grapheme==0.6.0
hijri-converter==2.2.3
holidays==0.13
httplib2==0.14.0
huggingface-hub==0.4.0
hyperlink==19.0.0
idna==2.8
incremental==16.10.1
install==1.3.5
itsdangerous==2.1.0
Jinja2==3.0.3
joblib==1.1.0
jsonpatch==1.28
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kiwisolver==1.3.2
korean-lunar-calendar==0.2.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
LunarCalendar==0.0.9
MarkupSafe==2.1.0
matplotlib==3.5.1
mysql-connector-python==8.0.28
netifaces==0.10.4
numpy==1.22.3
oauthlib==3.1.0
packaging==21.3
pandas==1.4.1
pexpect==4.6.0
Pillow==9.0.1
progress==1.6
progressbar2==4.0.0
prophet==1.0.1
protobuf==3.19.4
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.21
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMeeus==0.5.11
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.7
pyrsistent==0.17.3
pystan==2.19.1.1
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-utils==3.1.0
pytz==2021.3
PyYAML==5.3.1
regex==2022.3.2
requests==2.22.0
requests-unixsocket==0.2.0
sacremoses==0.0.47
scipy==1.8.0
SecretStorage==2.3.1
service-identity==18.1.0
setuptools-git==1.2
simplejson==3.16.0
six==1.14.0
sos==4.0
SpeechRecognition==3.8.1
ssh-import-id==5.10
systemd-python==234
tokenizers==0.11.6
torch==1.11.0+cpu
torchaudio==0.11.0+cpu
torchvision==0.12.0+cpu
tqdm==4.63.0
transformers==4.17.0
Twisted==18.9.0
typing-extensions==4.1.1
ubuntu-advantage-tools==20.3
ufw==0.36
ujson==5.1.0
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
Werkzeug==2.0.3
WTForms==3.0.1
zope.interface==4.7.1
